Add graphql schema expire time
We've seen instances where the graphql schema is not kept properly
sync'd with the DB schema. It's possible that the DB schema version is
not being properly updated  on one of the cluster's nodes? This adds a
fallback that will expire and rebuild the schema after a given time
regardless of the schema version.